refactor(app): extract useLegacyMigration hook (ngd)
Moves showLegacyMigrationModal state, the checkLegacyMigration effect, and
both import/skip handlers out of App.tsx into src/hooks/useLegacyMigration.ts.
Also fixes a pre-existing TS error: applies partial settings via immer draft
instead of passing Partial<AppSettings> directly.
